Вынес comments в отдельный reducer
