+++ src/reducer/data/movies-data/movies-data.js
@@ -7,18 +7,18 @@ const ActionTypes = {
   FETCH_MOVIES_DATA_ERROR: `FETCH_MOVIES_DATA_ERROR`,
   FETCH_TITLE_MOVIE: `FETCH_TITLE_MOVIE`,
   FETCH_TITLE_MOVIE_SUCCESS: `FETCH_TITLE_MOVIE_SUCCESS`,
-  FETCH_MOVIES_COMMENTS_DATA: `FETCH_MOVIES_COMMENTS_DATA`,
-  FETCH_MOVIES_COMMENTS_DATA_SUCCESS: `FETCH_MOVIES_COMMENTS_DATA_SUCCESS`,
+  // FETCH_MOVIES_COMMENTS_DATA: `FETCH_MOVIES_COMMENTS_DATA`,
+  // FETCH_MOVIES_COMMENTS_DATA_SUCCESS: `FETCH_MOVIES_COMMENTS_DATA_SUCCESS`,
 };
 
 const initialState = {
   loadingMovies: false,
   loadingMoviesError: false,
-  loadingComments: false,
+  // loadingComments: false,
   loadingTitleMovie: false,
   moviesList: [],
   titleMovie: {},
-  moviesComments: [],
+  // moviesComments: [],
 };
 
 const ActionCreator = {
@@ -32,13 +32,13 @@ const ActionCreator = {
   fetchMoviesDataError: () => ({
     type: ActionTypes.FETCH_MOVIES_DATA_ERROR,
   }),
-  fetchMoviesCommentsData: () => ({
-    type: ActionTypes.FETCH_MOVIES_COMMENTS_DATA,
-  }),
-  fetchMoviesCommentsDataSuccess: (moviesComments) => ({
-    type: ActionTypes.FETCH_MOVIES_COMMENTS_DATA_SUCCESS,
-    payload: moviesComments,
-  }),
+  // fetchMoviesCommentsData: () => ({
+  //   type: ActionTypes.FETCH_MOVIES_COMMENTS_DATA,
+  // }),
+  // fetchMoviesCommentsDataSuccess: (moviesComments) => ({
+  //   type: ActionTypes.FETCH_MOVIES_COMMENTS_DATA_SUCCESS,
+  //   payload: moviesComments,
+  // }),
   fetchTitleMovie: () => ({
     type: ActionTypes.FETCH_TITLE_MOVIE,
   }),
@@ -74,15 +74,15 @@ const reducer = (state = initialState, action) => {
       return extendObject(state, {
         loadingMoviesError: true,
       });
-    case ActionTypes.FETCH_MOVIES_COMMENTS_DATA:
-      return extendObject(state, {
-        loadingComments: true,
-      });
-    case ActionTypes.FETCH_MOVIES_COMMENTS_DATA_SUCCESS:
-      return extendObject(state, {
-        loadingComments: false,
-        moviesComments: action.payload
-      });
+    // case ActionTypes.FETCH_MOVIES_COMMENTS_DATA:
+    //   return extendObject(state, {
+    //     loadingComments: true,
+    //   });
+    // case ActionTypes.FETCH_MOVIES_COMMENTS_DATA_SUCCESS:
+    //   return extendObject(state, {
+    //     loadingComments: false,
+    //     moviesComments: action.payload
+    //   });
     case ActionTypes.FETCH_TITLE_MOVIE:
       return extendObject(state, {
         loadingTitleMovie: true,
