+++ src/reducer/data/movies-data/selectors.js
@@ -1,6 +1,6 @@
 import NameSpace from '../../name-space.js';
 
-const NAME_SPACE = NameSpace.DATA;
+const NAME_SPACE = NameSpace.MOVIE_DATA;
 
 export const getMovies = (state) => {
   return state[NAME_SPACE].moviesList;
@@ -10,9 +10,9 @@ export const getTitleMovie = (state) => {
   return state[NAME_SPACE].titleMovie;
 };
 
-export const getMoviesComments = (state) => {
-  return state[NAME_SPACE].moviesComments;
-};
+// export const getMoviesComments = (state) => {
+//   return state[NAME_SPACE].moviesComments;
+// };
 
 export const getMoviesLoadingStatus = (state) => {
   return state[NAME_SPACE].loadingMovies;
